refactor(bootstrap): extract platform layer to layer/platform
Phase 4.5c of the uc-bootstrap redesign. Move the OS-adapter construction
(SystemClipboardWiring, PlatformLayer, create_platform_layer,
noop_system_clipboard) out of assembly.rs into layer/platform.rs. The
struct was already fully pub, so no visibility change was needed. is_v2_blob
becomes pub(crate) (still in assembly until 4.5d) so the moved V2 blob purge
can call it; SystemClipboardWiring re-exports + deps.rs reference now source
from layer::platform; assembly drops the now-unused imports. assembly.rs:
1258 -> 1033 lines. No behavior change.
