+++ crates/uc-bootstrap/src/lib.rs
@@ -23,7 +23,7 @@ pub mod wiring;
 // `compose_event_context` 在 `wire_dependencies` 之后由各进程入口调用一次。
 pub use subsystem::analytics::compose_event_context;
 
-pub use assembly::{wire_dependencies, SystemClipboardWiring};
+pub use assembly::wire_dependencies;
 pub use entrypoint::daemon::{build_daemon_lifecycle, DaemonLifecycle};
 pub use entrypoint::non_gui::{
     build_app_facade_from_deps, build_cli_app_runtime, build_mobile_sync_facade,
@@ -31,6 +31,7 @@ pub use entrypoint::non_gui::{
     ClipboardRestoreAssembly,
 };
 pub use layer::paths::get_storage_paths;
+pub use layer::platform::SystemClipboardWiring;
 pub use observability::tracing::{init_tracing_subscriber, install_panic_logging_hook};
 pub use subsystem::blob_tasks::{spawn_blob_processing_tasks, BlobProcessingPorts};
 pub use subsystem::file_transfer::FileTransferLifecycle;
